Retrieve API keys as global environment variables
I found the method of retrieving keys from the config.yml file slightly confusing and inaccessible. If users would like to use the keys as global environment variables, they will be able to with the new method I added. The config.yml method is also still the first-most way of retreiving the keys, so this will not impact previous code.

I don't know as much about package documentation so I'm still getting the error Adding files missing in collate: retrieve_key.R

Apologies for not knowing how to fix this immediately, I'll be looking into it and hopefully adding another commit soon